requirements.txt
Django>=6.0
djangorestframework>=3.16
djangorestframework-simplejwt>=5.5
drf-nested-routers>=0.95
drf-spectacular>=0.29

psycopg2-binary>=2.9

pytest>=9.0
pytest-django>=4.11
